+++ backend/substrapp/tests/views/test_views_authentication.py
@@ -157,9 +157,9 @@ class AuthenticationTests(APITestCase):
 @override_settings(MEDIA_ROOT=MEDIA_ROOT)
 class TestLoginCase(APITestCase):
 
-    login_url = "/user/login/"
-    logout_url = "/user/logout/"
-    refresh_url = "/user/refresh/"
+    login_url = "/me/login/"
+    logout_url = "/me/logout/"
+    refresh_url = "/me/refresh/"
 
     username = "foo"
     password = "bar"