+++ backend/users/urls.py
@@ -7,10 +7,12 @@ from django.urls import path
 from rest_framework.routers import DefaultRouter
 
 # Create a router and register our viewsets with it.
+from users.views import AuthenticationViewSet
 from users.views import UserViewSet
 
 router = DefaultRouter()
-router.register(r"user", UserViewSet, basename="user")
+router.register(r"me", AuthenticationViewSet, basename="me")
+router.register(r"users", UserViewSet, basename="users")
 
 urlpatterns = [
     path("", include(router.urls)),