+++ backend/users/views/user.py
@@ -1,149 +1,211 @@
+import datetime
+
+import jwt
 from django.conf import settings
-from django.contrib.auth.models import User
+from django.contrib.auth import get_user_model
+from django.contrib.auth.password_validation import validate_password
+from django.core.exceptions import ValidationError as djangoValidationError
+from django.utils.encoding import force_str
+from jwt.exceptions import DecodeError
+from jwt.exceptions import ExpiredSignatureError
+from jwt.exceptions import InvalidTokenError
+from rest_framework import mixins
+from rest_framework import permissions
 from rest_framework import status
 from rest_framework.decorators import action
-from rest_framework.decorators import throttle_classes
-from rest_framework.permissions import AllowAny
-from rest_framework.response import Response
-from rest_framework.throttling import AnonRateThrottle
+from rest_framework.exceptions import ValidationError
+from rest_framework.filters import OrderingFilter
+from rest_framework.mixins import CreateModelMixin
 from rest_framework.viewsets import GenericViewSet
-from rest_framework_simplejwt.authentication import AUTH_HEADER_TYPES
-from rest_framework_simplejwt.exceptions import AuthenticationFailed
-from rest_framework_simplejwt.exceptions import InvalidToken
-from rest_framework_simplejwt.exceptions import TokenError
-from rest_framework_simplejwt.tokens import RefreshToken
 
-from libs.user_login_throttle import UserLoginThrottle
-from users.serializers import CustomTokenObtainPairSerializer
-from users.serializers import CustomTokenRefreshSerializer
+from libs.pagination import DefaultPageNumberPagination
+from substrapp.views.filters_utils import MatchFilter
+from substrapp.views.utils import ApiResponse
+from substrapp.views.utils import get_channel_name
+from users.models.user_channel import UserChannel
+from users.serializers.user import UserSerializer
 
 
-class UserViewSet(GenericViewSet):
-    queryset = User.objects.all()
-    serializer_class = CustomTokenObtainPairSerializer
+def _xor_secrets(secret1, secret2):
+    list = [chr(ord(s1) ^ ord(s2)) for s1, s2 in zip(secret1, secret2)]
+    return "".join(list)
 
-    www_authenticate_realm = "api"
 
-    permission_classes = [AllowAny]
+def _validate_channel(name):
+    if name not in settings.LEDGER_CHANNELS:
+        raise ValidationError({"channel": "Channel does not exist"})
 
-    def get_authenticate_header(self, request) -> str:
-        return f'{AUTH_HEADER_TYPES[0]} realm="{self.www_authenticate_realm}"'
 
-    @action(methods=["post"], detail=False)
-    @throttle_classes([AnonRateThrottle, UserLoginThrottle])
-    def login(self, request, *args, **kwargs):
-        serializer = self.get_serializer(data=request.data)
+def _validate_password(password, user):
+    try:
+        validate_password(password, user)
+    except djangoValidationError as err:
+        raise ValidationError(err.error_list)
 
-        try:
-            serializer.is_valid(raise_exception=True)
-        except AuthenticationFailed:
-            return Response({"message": "wrong username password"}, status=status.HTTP_401_UNAUTHORIZED)
-        except TokenError as e:
-            raise InvalidToken(e.args[0])
 
-        data = serializer.validated_data
+def _validate_role(role):
+    try:
+        role = UserChannel.Role[role]
+        return role
+    except KeyError:
+        raise ValidationError({"role": "Invalid role"})
 
-        refresh_token = data
-        access_token = data.access_token
 
-        access_expires = access_token.current_time + access_token.lifetime
-        refresh_expires = refresh_token.current_time + refresh_token.lifetime
+def _validate_token(token, secret):
+    try:
+        jwt.decode(token, secret, algorithms=[settings.RESET_JWT_SIGNATURE_ALGORITHM])
+    except (DecodeError, ExpiredSignatureError, InvalidTokenError) as err:
+        return {"is_valid": False, "message": err}
 
-        access_token_string = str(access_token)
-        header_payload = ".".join(access_token_string.split(".")[0:2])
-        signature = access_token_string.split(".")[2]
+    return {"is_valid": True, "message": ""}
 
-        response = Response(access_token.payload, status=status.HTTP_200_OK)
 
-        secure = not settings.DEBUG
+def _save_password(view, instance, password):
+    _validate_password(password, view.user_model(username=instance.username))
+    instance.set_password(password)
+    instance.save()
+    return UserSerializer(instance=instance).data
 
-        response.set_cookie(
-            "header.payload",
-            value=header_payload,
-            expires=access_expires,
-            secure=secure,
-            domain=settings.COMMON_HOST_DOMAIN,
+
+class IsAdminOrReadOnly(permissions.BasePermission):
+    def has_permission(self, request, view):
+        if request.method in permissions.SAFE_METHODS:
+            return True
+
+        return request.user.channel.role == UserChannel.Role.ADMIN
+
+
+class IsSelf(permissions.BasePermission):
+    def has_permission(self, request, view):
+        user = view.get_object()
+        if not user.is_authenticated:
+            return False
+
+        return user.id == request.user.id
+
+
+class UserViewSet(
+    GenericViewSet,
+    CreateModelMixin,
+    mixins.ListModelMixin,
+    mixins.RetrieveModelMixin,
+    mixins.DestroyModelMixin,
+):
+    user_model = get_user_model()
+    serializer_class = UserSerializer
+    pagination_class = DefaultPageNumberPagination
+    ordering_fields = ["username"]
+    ordering = ["username"]
+    filter_backends = [OrderingFilter, MatchFilter]
+    lookup_field = "username"
+    permission_classes = [permissions.IsAuthenticated, IsAdminOrReadOnly]
+    search_fields = ["username"]
+
+    def get_queryset(self):
+        channel = get_channel_name(self.request)
+        return self.user_model.objects.filter(channel__channel_name=channel)
+
+    def create(self, request, *args, **kwargs):
+        username = request.data.get("username")
+        password = request.data.get("password")
+        channel = get_channel_name(self.request)
+        role = request.data.get("role")
+
+        _validate_channel(channel)
+        _validate_password(password, self.user_model(username=username))
+
+        channel_data = {"channel_name": channel}
+        if role:
+            channel_data["role"] = _validate_role(role)
+
+        user = self.user_model.objects.create_user(username=username, password=password)
+
+        channel_data["user"] = user
+
+        UserChannel.objects.create(**channel_data)
+
+        user.refresh_from_db()
+        data = UserSerializer(instance=user).data
+
+        return ApiResponse(data=data, status=status.HTTP_201_CREATED, headers=self.get_success_headers(data))
+
+    def update(self, request, *args, **kwargs):
+        instance = self.get_object()
+        role = request.data.get("role")
+
+        if role:
+            role = _validate_role(role)
+            instance.channel.role = role
+            instance.channel.save()
+
+        instance.save()
+        data = UserSerializer(instance=instance).data
+
+        return ApiResponse(data=data, status=status.HTTP_200_OK, headers=self.get_success_headers({}))
+
+    @action(methods=["put"], detail=True, permission_classes=[IsSelf])
+    def password(self, request, *args, **kwargs):
+        """Allows an authenticated user to modify his own password"""
+        instance = self.get_object()
+        password = request.data.get("password")
+
+        if password:
+            data = _save_password(self, instance, password)
+            return ApiResponse(data=data, status=status.HTTP_200_OK, headers=self.get_success_headers({}))
+
+        return ApiResponse(data={"message": "missing password in the request"}, status=status.HTTP_400_BAD_REQUEST)
+
+    @action(methods=["post"], detail=True, permission_classes=[permissions.AllowAny], url_name="set-password")
+    def set_password(self, request, *args, **kwargs):
+        """Allows unauthenticated user to set new password if valid reset token is provided"""
+        token = request.data.get("token")
+        new_password = request.data.get("password")
+
+        username = kwargs.get("username")
+        instance = self.user_model.objects.get(username=username)
+
+        secret = _xor_secrets(instance.password, force_str(settings.SECRET_KEY))
+        token_validation = _validate_token(token, secret)
+
+        if token_validation.get("is_valid"):
+            data = _save_password(self, instance, new_password)
+            return ApiResponse(data=data, status=status.HTTP_200_OK, headers=self.get_success_headers({}))
+
+        return ApiResponse(
+            data={"message": "must provide a valid token to set a new password"},
+            status=status.HTTP_401_UNAUTHORIZED,
         )
-        response.set_cookie(
-            "signature", value=signature, httponly=True, secure=secure, domain=settings.COMMON_HOST_DOMAIN
-        )
-        response.set_cookie(
-            "refresh",
-            value=str(refresh_token),
-            expires=refresh_expires,
-            httponly=True,
-            secure=secure,
-            domain=settings.COMMON_HOST_DOMAIN,
-        )
 
-        return response
+    @action(detail=True, permission_classes=[permissions.AllowAny], url_name="verify-token")
+    def verify_token(self, request, *args, **kwargs):
+        """Return 200 if reset token is valid 401 otherwise. Accepts unauthenticated request"""
+        token = request.query_params.get("token", None)
 
-    @action(methods=["post"], detail=False)
-    def refresh(self, request, *args, **kwargs):
-        serializer = CustomTokenRefreshSerializer(data=request.data, context=self.get_serializer_context())
+        username = kwargs.get("username")
+        instance = self.user_model.objects.get(username=username)
 
-        try:
-            serializer.is_valid(raise_exception=True)
-        except AuthenticationFailed:
-            return Response({"message": "wrong username password"}, status=status.HTTP_401_UNAUTHORIZED)
-        except TokenError as e:
-            raise InvalidToken(e.args[0])
+        secret = _xor_secrets(instance.password, force_str(settings.SECRET_KEY))
+        token_validation = _validate_token(token, secret)
+        if token_validation.get("is_valid"):
+            return ApiResponse(data={}, status=status.HTTP_200_OK, headers=self.get_success_headers({}))
 
-        data = serializer.validated_data
-
-        refresh_token = data
-        access_token = data.access_token
-
-        access_expires = access_token.current_time + access_token.lifetime
-        refresh_expires = refresh_token.current_time + refresh_token.lifetime
-
-        access_token_string = str(access_token)
-        header_payload = ".".join(access_token_string.split(".")[0:2])
-        signature = access_token_string.split(".")[2]
-
-        response = Response(access_token.payload, status=status.HTTP_200_OK)
-
-        secure = not settings.DEBUG
-
-        response.set_cookie(
-            "header.payload",
-            value=header_payload,
-            expires=access_expires,
-            secure=secure,
-            domain=settings.COMMON_HOST_DOMAIN,
-        )
-        response.set_cookie(
-            "signature", value=signature, httponly=True, secure=secure, domain=settings.COMMON_HOST_DOMAIN
-        )
-        response.set_cookie(
-            "refresh",
-            value=str(refresh_token),
-            expires=refresh_expires,
-            httponly=True,
-            secure=secure,
-            domain=settings.COMMON_HOST_DOMAIN,
+        return ApiResponse(
+            data={"message": f"token not valid: {token_validation.get('message')}"},
+            status=status.HTTP_401_UNAUTHORIZED,
         )
 
-        return response
+    @action(methods=["post"], detail=True, url_path="reset_password", url_name="reset-password")
+    def generate_reset_password_token(self, request, *args, **kwargs):
+        """Returns reset password token. Restricted to Admin request"""
+        instance = self.get_object()
+        secret = _xor_secrets(instance.password, force_str(settings.SECRET_KEY))
 
-    @action(detail=False)
-    def logout(self, request, *args, **kwargs):
+        jwt_token = jwt.encode(
+            payload={"exp": datetime.datetime.now() + datetime.timedelta(days=7)},
+            key=secret,
+            algorithm=settings.RESET_JWT_SIGNATURE_ALGORITHM,
+        )
 
-        # Blacklist jwt token at logout fetched from cookie
-        if "refresh" in request.COOKIES:
-            refresh = RefreshToken(request.COOKIES["refresh"])
-            try:
-                # Attempt to blacklist the fetched refresh token
-                refresh.blacklist()
-            except AttributeError:
-                # If blacklist app not installed, `blacklist` method will
-                # not be present
-                pass
+        data = {"reset_password_token": jwt_token}
 
-        response = Response({}, status=status.HTTP_200_OK)
-
-        response.delete_cookie("header.payload", domain=settings.COMMON_HOST_DOMAIN)
-        response.delete_cookie("signature", domain=settings.COMMON_HOST_DOMAIN)
-        response.delete_cookie("refresh", domain=settings.COMMON_HOST_DOMAIN)
-
-        return response
+        return ApiResponse(data=data, status=status.HTTP_200_OK, headers=self.get_success_headers({}))