+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - chore: rename connect-tools to substra-tools
 - stop generating algo method to execute from task category as it is now passed within substra cli
 
+### Added
+
+- BREAKING: allow registration of users with role from the API and reset password mechanism
+
 ## [0.28.0] 2022-08-29
 
 ### Added